pyproject.toml
[tool.poetry]
name = "souris_gui_test"
version = "0.1.0"
description = "Test gui for souris app"
authors = ["MooSee2 <52461645+MooSee2@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.9.12"
dash = "^2.14.2"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.0"
dash-bootstrap-components = "^1.5.0"
dash-mantine-components = "^0.12.1"
sqlalchemy = "^2.0.28"
pandas-stubs = "^2.2.1.240316"
flask-sqlalchemy = "^3.1.1"
pymysql = "^1.1.0"
flask-caching = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 999
